Repair SliceHeader.cpp Warning

+++ SliceHeader.cpp
@@ -357,7 +357,7 @@ void SliceHeader::pred_weight_table() {
     }
   }
 
-  for (int i = 0; i <= num_ref_idx_l0_active_minus1; ++i) {
+  for (uint32_t i = 0; i <= num_ref_idx_l0_active_minus1; ++i) {
     cout << "\t\t前参考帧列表亮度权重:" << luma_weight_l0[i] << endl;
     cout << "\t\t前参考帧列表亮度权重:" << luma_offset_l0[i] << endl;
     cout << "\t\t前参考帧列表色度权重:" << chroma_weight_l0[i][0] << endl;
@@ -470,51 +470,51 @@ int SliceHeader::set_scaling_lists_values() {
             // 参照 Table 7-2 Scaling list fall-back rule A
             if (i == 0)
               memcpy(ScalingList4x4[i], Default_4x4_Intra,
-                     sizeof(int32_t) * 16);
+                     sizeof(uint8_t) * 16);
             else if (i == 3)
               memcpy(ScalingList4x4[i], Default_4x4_Inter,
-                     sizeof(int32_t) * 16);
+                     sizeof(uint8_t) * 16);
             else
               memcpy(ScalingList4x4[i], ScalingList4x4[i - 1],
-                     sizeof(int32_t) * 16);
+                     sizeof(uint8_t) * 16);
           } else {
             if (m_pps->UseDefaultScalingMatrix4x4Flag[i] == 1) {
               if (i < 3)
                 memcpy(ScalingList4x4[i], Default_4x4_Intra,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
               else
                 memcpy(ScalingList4x4[i], Default_4x4_Inter,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
 
             } else
               // 采用编码器传送过来的量化系数的缩放值
               memcpy(ScalingList4x4[i], ScalingList4x4[i],
-                     sizeof(int32_t) * 16);
+                     sizeof(uint8_t) * 16);
           }
         } else {
           if (m_sps->seq_scaling_list_present_flag[i] == 0) {
             // 参照 Table 7-2 Scaling list fall-back rule A
             if (i == 6)
               memcpy(ScalingList8x8[i - 6], Default_8x8_Intra,
-                     sizeof(int32_t) * 64);
+                     sizeof(uint8_t) * 64);
             else if (i == 7)
               memcpy(ScalingList8x8[i - 6], Default_8x8_Inter,
-                     sizeof(int32_t) * 64);
+                     sizeof(uint8_t) * 64);
             else
               memcpy(ScalingList8x8[i - 6], ScalingList8x8[i - 8],
-                     sizeof(int32_t) * 64);
+                     sizeof(uint8_t) * 64);
 
           } else {
             if (m_pps->UseDefaultScalingMatrix8x8Flag[i - 6] == 1) {
               if (i == 6 || i == 8 || i == 10)
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Intra,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
               else
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Inter,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
             } else
               memcpy(ScalingList8x8[i - 6], ScalingList8x8[i - 6],
-                     sizeof(int32_t) * 64);
+                     sizeof(uint8_t) * 64);
             // 采用编码器传送过来的量化系数的缩放值
           }
         }
@@ -531,30 +531,30 @@ int SliceHeader::set_scaling_lists_values() {
             if (i == 0) {
               if (m_sps->seq_scaling_matrix_present_flag == 0) {
                 memcpy(ScalingList4x4[i], Default_4x4_Intra,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
               }
             } else if (i == 3) {
               if (m_sps->seq_scaling_matrix_present_flag == 0) {
                 memcpy(ScalingList4x4[i], Default_4x4_Inter,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
               }
             } else {
               memcpy(ScalingList4x4[i], ScalingList4x4[i - 1],
-                     sizeof(int32_t) * 16);
+                     sizeof(uint8_t) * 16);
             }
           } else {
             if (m_pps->UseDefaultScalingMatrix4x4Flag[i] == 1) {
               if (i < 3) {
                 memcpy(ScalingList4x4[i], Default_4x4_Intra,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
               } else // if (i >= 3)
               {
                 memcpy(ScalingList4x4[i], Default_4x4_Inter,
-                       sizeof(int32_t) * 16);
+                       sizeof(uint8_t) * 16);
               }
             } else {
               memcpy(ScalingList4x4[i], ScalingList4x4[i],
-                     sizeof(int32_t) *
+                     sizeof(uint8_t) *
                          16); // 采用编码器传送过来的量化系数的缩放值
             }
           }
@@ -566,30 +566,30 @@ int SliceHeader::set_scaling_lists_values() {
             if (i == 6) {
               if (m_sps->seq_scaling_matrix_present_flag == 0) {
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Intra,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
               }
             } else if (i == 7) {
               if (m_sps->seq_scaling_matrix_present_flag == 0) {
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Inter,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
               }
             } else {
               memcpy(ScalingList8x8[i - 6], ScalingList8x8[i - 8],
-                     sizeof(int32_t) * 64);
+                     sizeof(uint8_t) * 64);
             }
           } else {
             if (m_pps->UseDefaultScalingMatrix8x8Flag[i - 6] == 1) {
               if (i == 6 || i == 8 || i == 10) {
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Intra,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
               } else {
                 memcpy(ScalingList8x8[i - 6], Default_8x8_Inter,
-                       sizeof(int32_t) * 64);
+                       sizeof(uint8_t) * 64);
               }
             } else {
               memcpy(ScalingList8x8[i - 6], ScalingList8x8[i - 6],
-                     sizeof(int32_t) *
-                         64); // 采用编码器传送过来的量化系数的缩放值
+                     sizeof(uint8_t) * 64);
+              // 采用编码器传送过来的量化系数的缩放值
             }
           }
         }
